requirements.txt
logging>=0.4.96
paho-mqtt>=1.4.0
pyserial>=3.4
